ci: switch Go version to latest release; lint
Previously, we used "master" version of Go on Travis CI.
Apparently, this uses a tip version and not the latest release.

So, use a fixed version.
The downside is we'll need to update it with each new Go release,
but it's still better than building release binaries using
a tip version.